Split Rope and Jedi completions
